requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0
Pillow>=10.2.0
numpy>=1.26.0
